==> Pandas/1-pd-series.ipynb <==
import pandas as pd
import numpy as np

## Ejercicio 1

 ¿Qué diferencia hay entre una Serie y un DataFrame?

Ambas son estructuras de datos de Pandas, pero las series son estructuras en una dimensión y los DataFrame son estructuras en dos dimensiones (tablas).

## Ejercicio 2 : Series

Momento de trabajar con Series. Contestad a las siguientes preguntas:

Crea dos listas:
- Una que se llame "valores" que tenga 7 números aleatorios.
- Otra que se llame "indices" que tenga 7 strings que se llamen como los días de la semana.

#Hemos consultado: https://stackoverflow.com/a/25174767 para obtener una muestra con números no duplicados
valores = np.random.choice(range(20), 7, replace=False)
valores

indices = ["lunes", "martes", "miercoles", "jueves", "viernes", "sabado", "domingo"]

- A partir de estas dos listas cread una Serie donde la variable "valores" sean los valores de la Serie y la variable indices sean los índices de la Serie.

serie1 = pd.Series(indices,valores)
serie1

- A partir de las dos listas anteriores cread un diccionario usando un bucle for. Para hacer esto tenemos el método zip que nos permite iterar por dos listas a la vez. 

def diccionizador(val,ind):
    diccionario = dict(zip(val, ind))
    return diccionario

        

diccionizador(valores,indices)

diccionario = diccionizador(indices,valores)
diccionario

- Crea una Serie con el diccionario creado en el punto anterior

series2 = pd.Series(diccionario)
series2

- ¿Qué forma tiene esta última Serie?

series2.shape
print(f"Nuestra serie tiene una forma de {series2.shape} filas")

- ¿Qué tipo de datos tiene?

series2.dtypes

- ¿Qué tamaño tiene?

series2.size

- ¿Cuáles son los índices de la Serie?

series2.index

- ¿Y los values?

series2.values

- Extraed el valor para el índice "lunes"

series2["lunes"]

- Extraed el valor para el índice "lunes", "martes" y "miércoles".

series2["lunes":"miercoles"]

- Extraed el valor para el índice "lunes" y "domingo".

series2[[0,-1]]

## Ejercicio 3 : DataFrames
- Cread un array de dos dimensiones con 4 filas y 3 columnas.

array_ej3 = np.random.randint(0,50,(4,3))
array_ej3

- Convertid el array en un DataFrame donde las columnas se llamen "España", "Francia" e "Italia".

df_ejer3 = pd.DataFrame(array_ej3, columns = ['España', 'Francia', 'Italia'])
df_ejer3

- Descargad el archivo "medallas" y cargad el DataFrame con la siguiente línea de código:

df_medallas = pd.read_csv("medallas.csv", index_col=["País"])
df_medallas

 NOTA: Para evitar problemas guardad el archivo medallas.csv a la misma altura (en la misma carpeta) que el jupyter donde estamos trabajando.

- Usando el loc extraed todas las columnas de la fila de Tailandia.

df_medallas.loc["Tailandia",]

- Usando el loc extraed todas las columnas de China e Irán.

China = df_medallas.loc["China",]
Irán = df_medallas.loc["Irán",]
print(China,Irán)

- Usando el loc extraed solo las columnas de "Oro", "Plata" y "Bronce" de China e Irán.

colum_China_Iran =df_medallas.loc["China",["Oro","Plata","Bronce"]], df_medallas.loc["Irán",["Oro","Plata","Bronce"]]
colum_China_Iran

- Usando el iloc extraed la información de la India. Devolved todas las columnas.

df_medallas.iloc[5,:]

- Usando el iloc extraed todas las columnas de los paises que estén en posición impar. La primera que nos tendría que salir es Japón y la última Corea del Norte.

df_medallas.iloc[1::2,:]

- Utilizando el código del ejercicio anterior, seleccionad unicamente las columnas "Oro", "Plata" y "Bronce".

df_medallas.iloc[1::2,1:]

Contestad a las siguientes preguntas:

- ¿Qué país consiguió más de 200 medallas de oro?

df_medallas.loc[df_medallas.Oro > 200, :]

- ¿Qué país consiguió más de 200 medallas de oro y más de 1000 de plata?

df_medallas.iloc[list((df_medallas.Oro > 200) & (df_medallas.Plata > 1000)), :]

- ¿Qué país consiguió más de 200 medallas de oro o más de 700 de bronce?

df_medallas.iloc[list((df_medallas.Oro > 200) | (df_medallas.Bronce > 700)), :]

- Como vemos hay una columna que se llama Unnamed: 0 que no nos interesa. Eliminad esa columna.

df_medallas.drop(['Unnamed:0'], axis=1)

df_medallas.drop(df_medallas["Unnamed:0"])

- Cread una nueva columma que se llame "total" que sea el resultado de la suma de todas las medallas obtenidas por cada país.

df_medallas.assign(total=df_medallas["Oro"] + df_medallas["Plata"] + df_medallas["Bronce"])
