==> Pandas/2-pd-car-y-guar.ipynb <==
import pandas as pd
import pickle
import numpy as np

## Ejercicio 1
Id cargando todos los ficheros en vuestro jupyter.

- Archivo .JSON

df_json = pd.read_json('../pandas/quejas.json')
df_json.head()

df_json2 = df_json['data'].apply(pd.Series)
df_json2.head()

- Archivo CSV:

df_csv = pd.read_csv('../pandas/quejas.csv', encoding = 'latin-1', sep = ';').head(5)
df_csv

df_xlsx =

- Archivo PICKLE

df_pickle = pd.read_pickle('../pandas/algas.pickle')
df_pickle.head()

- Archivo .XLSX

df_xlsx = pd.read_excel('../pandas/quejas2.xlsx')
df_xlsx.head()

## Ejercicio 2
Para poner en práctica el read_clipboard, elegid alguna tabla que os podaís encontrar en google y usad dicho método para traer los datos al jupyter notebook. Una vez que los hayáis cargado, guardad esos datos en los siguientes formatos:  
        - .csv  
        - .excel  
        - .json  
        - .pickle  

#la tabla es sobre items clave del videojuego Persona 5 Royal.
#https://megamitensei.fandom.com/wiki/List_of_Persona_5_Royal_Items
df_clipboard = pd.read_clipboard(sep = '\t')
df_clipboard.head(10)

Una vez cargada la tabla, guardamos los datos:

- En CSV:

df_clipboard.to_csv('../pandas/tabla_persona.csv')

- En Excel:

df_clipboard.to_excel('../pandas/tabla_persona.xlsx')

- En Pickle:

df_clipboard.to_pickle('../pandas/tabla_persona.pkl')

- En JSON:

df_clipboard.to_json('../pandas/tabla_persona.json')

- En Parquet:

df_clipboard.to_parquet('../pandas/tabla_persona.parquet')
