==> Pandas/3-pd-metod.ipynb <==
import pandas as pd

db_sharks = pd.read_csv('../Pandas/Archivos/attacks.csv', encoding='UTF-8')
db_sharks.head(2)

db_sharks.info()

### Ejercicio 2
Haced una exploración básica del DataFrame:

* ¿Cuántas filas y columnas tiene el DataFrame?

print(f'El DataFrame tiene {db_sharks.shape[0]} filas y {db_sharks.shape[1]} columnas.')

* Mostrad las primeras cinco filas y las últimas cinco filas del DataFrame.

*   Primeras cinco filas:

#primeras cinco filas. no sería necesario especificar el 5 ya que, por defecto, head() devuelve las 5 primeras filas.
db_sharks.head(5)

* Primeras últimas filas:

#al igual que head, tail devuelve por defecto las 5 primeras filas:
db_sharks.tail()

* Devuelve dos filas al azar:

db_sharks.sample(2)

* ¿Cuál es el nombre de las columnas del DataFrame?

db_sharks.columns

* ¿Cuáles son los principales estádisticos del DataFrame? Mostrad por separado las variables categóricas y las variables numéricas.

#variables numéricas
db_sharks.describe()

#variables categóricas
db_sharks.describe(include=object)

        Nos pasa lo siguiente que no sabemos como resolver: Sobre los originales se guardan los cambios aplicados con el .strip() pero no con las copias. Por ello, hemos optado por quitar primero los espacios y luego hacer la copia.

#vamos a quitar primero los espacios que hay después de cada nombre de columna:
#https://stackoverflow.com/a/41476181
db_sharks.columns = db_sharks.columns.str.strip()

#pendiente hacer un regex con los espacios al final de los nombres de las columnas para que los reemplace por "".
db_sharks.columns = db_sharks.columns.str.replace(" ", "_").str.lower()
db_sharks.columns

db_sharks2 = db_sharks.copy()
db_sharks2.columns

* Eliminad las columnas de Unnamed: 22 y Unnamed: 23, Case Number.1, Case Number.2, original order, Investigator or Source, pdf, href formula.

db_sharks2.drop(['Unnamed:_0', 'Investigator_or_Source', 'Case_Number.1', 'Case_Number.2', 'original_order', 'Unnamed:_22',
       'Unnamed:_23'], inplace=True, axis=1)
db_sharks2.columns

* Estableced la columna 'case_number' como índice del dataframe:

db_sharks2.set_index('case_number', inplace=True)
db_sharks2.head()

### Ejercicio 3.
Guarda el csv que no tiene esas columnas para seguir usándolo mañana.

* Por último, ¡guardamos los datos en un nuevo csv! Le cambiamos el nombre a esta versión para tener trazabilidad de los cambios aplicados en el DataFrame:

db_csv = db_sharks2.to_csv('../Pandas/attacks_1_2.csv')
db_csv
