==> Pandas/4-pd-fil-datos.ipynb <==
import pandas as pd
import numpy as np
import re

### Ejercicio 1

Cargad el csv que guardasteis ayer.
Antes de seguir con las preguntas, este DataFrame está terriblemente sucio. En las próximas sesiones lo iremos limpiando. Para evitar posibles errores con los nulos vamos a eliminar todos los nulos del DataFrame.
Para ello vamos a utilizar el método .dropna() de pandas. 
`df.dropna(inplace = True)`
¿Cuántas filas se han eliminado con esta línea de código?

#primero abrimos la base de datos
db_sharks = pd.read_csv('../Archivos/attacks_1_2.csv')
db_sharks.head(5)

db_sharks2 = db_sharks.copy()
db_sharks2.head()

* Antes de manipular la BBDD, vamos a hacer una exploración preliminar del dataframe:

#información básica sobre el dataframe
db_sharks2.info()

#Conteo de filas
filas_antes = db_sharks2.shape[0]
print(f'Antes de borrar los nulos, hay un total de {filas_antes} casos')

#Eliminamos todos los nulos del dataframe
db_sharks2.dropna(inplace = True)

#conteo de filas después de borrar los valores nulos
filas_despues = db_sharks2.shape[0]
print(f'Después de borrar los nulos, hay un total de {db_sharks2.shape[0]} casos')

#calculamos el total de filas que se han borrado:
print(f"Con la eliminación de los casos nulos, se han borrado un total de {(filas_antes - filas_despues)} filas.")

### Ejercicio 2
¿Cuál es el número de ataques que se han producido en Estados Unidos?

#volvemos a abrir el df para tenerlo presente:
db_sharks2.head(2)

db_usa = db_sharks2[db_sharks2["country"] == "USA"]
print(f"Ha habido un total de {db_usa.country.count()} casos de ataques de tiburones en USA.")

### Ejercicio 3
¿Y en España?

db_spain = db_sharks2[db_sharks2["country"] == "SPAIN"]
print(f"Ha habido un total de {db_spain.country.count()} casos de ataques de tiburones en España.")

* 3.1 ¿En qué regiones se han producido los ataques (columna location) ?

db_spain['location'].unique()

* 3.2 ¿Cuántos ataques se han producido en cada área (columna area)?

db_spain['location'].value_counts()

### Ejercicio 4

Extraed un DataFrame nuevo con los ataques que se han producido en Australia y Estados Unidos. 📌 Nota Este ejercicio se puede resolver usando dos condiciones de filtrado o usando el método isin(). Hacedlo con ambas herramientas.

* Nuevo DF con condiciones de filtrado:

db_usa_aus = db_sharks2[(db_sharks2['country'] == 'USA') | (db_sharks2['country'] == 'AUSTRALIA')]
db_usa_aus.head(3)

#confirmamos que el nuevo df solo registra los resultados de los países que nos interesan:
db_usa_aus["country"].unique()

* Nuevo DF con isin():

db_usa_aus = db_sharks2[db_sharks2['country'].isin(['USA', 'AUSTRALIA'])]
db_usa_aus.head(3)

#confirmamos, de nuevo, que con el mismo método el nuevo df solo registra los resultados de los países que nos interesan:
db_usa_aus["country"].unique()

* 4.1 ¿Cuántos ataques se han producido en cada país?

db_usa_aus['country'].value_counts()

* 4.2 ¿Ha habido ataques en Australia a gente que estuviera remando (lo encontrareis como Paddling en el DataFrame en la columna activity)?

aus_rem = db_usa_aus[(db_usa_aus['country'] == 'AUSTRALIA') & (db_usa_aus['activity'] == 'Paddling')]
aus_rem.head(3)

        Vemos que con las condiciones aplicadas, el filtro no nos devuelve ningún caso. Igualmente, vamos a confirmarlo contando el número de casos:

len(aus_rem)
print(f'Hay un total de {len(aus_rem)} casos en los que haya habido ataques en Australia a gente que estuviera remando') 

* 4.3 ¿Cuántos ataques se han producido en Australia antes del 2000?

cond1 = db_usa_aus['country'] == 'AUSTRALIA'
cond2 = db_usa_aus['year'] < 2000
db43= db_usa_aus[(cond1) & (cond2)]
db43.head(3)

len(db43)
print(f"Ha habido un total de {len(db43)} ataques de tiburón en Australia antes del 2000")

* 4.4 ¿Y después del 2000?

cond1 = db_usa_aus['country'] == 'AUSTRALIA'
cond2 = db_usa_aus['year'] > 2000
db44= db_usa_aus[(cond1) & (cond2)]
db44.head(2)

#de nuevo, vamos a contar los casos que cumplen nuestras condiciones:
db_usa_aus[(cond1) & (cond2)].shape[0]
print(f"Ha habido un total de {db_usa_aus[(cond1) & (cond2)].shape[0]} ataques de tiburón en Australia después del 2000")

### Ejercicio 5
Extraed unicamente los ataques a mujeres.

at_muj = db_sharks2[db_sharks2['sex'] == 'F']
at_muj.head(2)

#comprobamos el número de casos:
at_muj.shape[0]
print(f'Ha habido un total de {at_muj.shape[0]} ataques a mujeres')

### Ejercicio 6
¿Cuántas mujeres fueron atacadas en España?

at_muj_esp = db_spain[db_spain['sex'] == 'F']
at_muj_esp.head(2)

at_muj_esp.shape[0]
print(f'Ha habido {at_muj_esp.shape[0]} ataque a mujeres en España.')

### Ejercicio 7
Extraed los ataques que se hayan producido en el año 2000, 1997 y 2006. Utilizad el método isin() para resolver este ejercicio

db_sharks2[db_sharks2['year'].isin([1997, 2000, 2006])].head(3)

### Ejercicio 8
Si extraemos los valores únicos de la columna species nos podemos dar cuenta que está un poco sucia y extraer informacion útil de la misma se puede convertir en algo difícil. En este ejercicio tendréis que extraer los ataques que han ocurrido por tiburón blanco (lo encontraremos en la columna species como algo que contenga white o White). Usad el método str.contains() para resolver este ejercicio.

tibblanco = db_sharks2[db_sharks2["species"].str.contains('[w|W]hite')]
tibblanco.head(3)

#vamos a comprobar las categorías:
tibblanco['species'].unique()

        Vemos que hay muchos casos que incluyen 'white shark'. Comprobamos que el código ha funcionado en el primer valor de la línea 1 (posición 0 del array), y en el primer valor de la segunda línea (posición 4 del array).

### Ejercicio 9
Extraed ahora los ataques por tiburón blanco o tigre. Utilizad el método str.contains() para resolver este ejercicio.

tibbltig = db_sharks2[db_sharks2["species"].str.contains('[w|W]hite | [t|T]iger')]
tibbltig.head(3)

        Vamos a comprobar que nuestro filtro ha funcionado:

tibbltig['species'].unique()

        Vemos que los valores de tiger son muy difíciles de ver.
        Para ello, vamos a confirmar que están de la siguiente manera:

tiger = tibbltig[tibbltig['species'].str.contains('[t|T]iger')] #almacenamos el filtro en una variable
tiger['species'].unique() #consultamos los valores únicos de la variable que hemos creado

### Ejercicio 10
Utilizando la columna de year cread una nueva columna llamada siglo que tiene siglo-XX o siglo-XXI. Donde el año sea menor que 2000 pondremos siglo-XX y el resto de los años siglo-XXI.


db_sharks3 = db_sharks2.copy() #como se trata de un cambio importante en el csv, vamos a trabajar sobre una copia
db_sharks3.head(2)

#creamos la columna con los filtros:
siglos = [(db_sharks3['year'] < 1900), (db_sharks3['year'] < 2000), (db_sharks3['year'] >= 2000)]
valores = ['siglo-XIX', 'siglo-XX', 'siglo-XXI']

db_sharks3['siglo'] = np.select(siglos, valores)

#comprobamos que los valores se han introducido correctamente:
db_sharks3['siglo'].unique()

        Vamos a comprobar que nuestra columna siglo ha filtrado correctamente los valores de la columna year. Para ello buscamos el valor mínimo de year y su posición:

db_sharks3.loc[db_sharks3['year'] == db_sharks3['year'].min()]

        Vemos que en la columna 'siglo' en este caso aparece 'siglo-XIX', que es el valor que queríamos que apareciera. Ha funcionado :-)

## Ejercicio 11
Guardad como un archivo csv nuevo. Lo usaremos en el pair programming de mañana.

db_sharks3 = db_sharks3.to_csv('../Archivos/attacks_1_3.csv', index=False)
