==> Pandas/5-pd-union-dat.ipynb <==
### Ejercicio 1.
Cargad los tres ficheros nombrados anteriormente y los exploramos para familizarizarnos con ellos y saber que información tienen en común y debatir con tu compañera como podríamos juntar toda la información.

import pandas as pd
import numpy as np

attacks = pd.read_csv('../Archivos/attacks_1_3.csv')
attadicional = pd.read_csv('../Archivos/attacks_adicional.csv')
atthist = pd.read_csv('../Archivos/attacks_historico.csv')

attacks.head(3)

attadicional.head(3)

atthist.head(3)

attacks.info()

attadicional.info()

atthist.info()

        Podemos observar que attacks_historico.csv (atthist) las columnas son las mismas que en nuestro archivo original de attacks.csv antes de la depuración.

### Ejercicio 2.
Chequead que los nombres de las columnas ver si se llaman iguales entre dataframes . ¿Están todas en minúsculas? ¿Tienen espacios? En caso de que no sean todas iguales unificad el nombre de las columnas.

#vamos a crear copias para cada archivo:
attacks2 = attacks.copy()
atthist2 = atthist.copy()
attadic2 = attadicional.copy()

        Es necesario homogeneizar los formatos de los nombres de las columnas (están capitalizadas) y quitar los espacios para que se parezca lo máximo posible a nuestro csv attacks. También hay que quitar las columnas y los nulos que borramos en ejercicios anteriores.

* Limpieza de atthist (attacks_historicos.csv):

#quitamos los espacios delante y detrás de los nombres:
atthist.columns = atthist.columns.str.strip()
#quitamos los espacios y los reemplazamos por '_':
atthist.columns = atthist.columns.str.replace(" ", "_").str.lower()
atthist.columns
#Nos pasa lo siguiente que no sabemos como resolver: Sobre los originales se guardan los cambios aplicados con el .strip() pero no con las copias

#borramos las columnas que no nos hacen falta
atthist2.drop(['unnamed:_0', 'investigator_or_source', 'case_number.1', 'case_number.2', 'original_order', 'unnamed:_22',
       'unnamed:_23'], inplace=True, axis=1)
atthist2.columns

* Limpieza de attadic2 (attacks_adicional.csv)

#homogeneizamos el formato:
attadic2.columns = attadic2.columns.str.strip().str.replace(" ", "_").str.lower()
#borramos la columna innecesaria:
attadic2.drop(['unnamed:_0'], inplace=True, axis=1)
attadic2.columns

### Ejercicio 3.
En este momento nos habremos dado cuenta que el DataFrame que generamos en el pair programming Pandas IV es similar al DataFrame de los datos históricos. Si recordamos, eliminamos algunas columnas en la sesión III que no eran útiles, pero esas columnas están todavía en el histórico. Si no fuera por eso podríamos unir esos dataframes. Realizad los cambios que creais necesarios para que los dataframes sean iguales y luego juntadlos usando el mejor método.

        Sabemos que la única columna que tiene ahora attacks2 que no tiene attthist2 es la columna de 'siglos'. Creamos la columna siglos en atthist2:

atthist2.columns

atthist2['year'].unique()

#creamos la columna con los filtros:
        #[(db_sharks3['year'] < 1900), (db_sharks3['year'] < 2000), (db_sharks3['year'] >= 2000)]
siglos = [(atthist2['year'] == 1800), (atthist2['year'] < 1800), (atthist2['year'] < 1700), (atthist2['year'] < 1600)]
valores = ['siglo XIX', 'siglo-XVIII', 'siglo-XVII', 'siglo-XVI o anterior']

atthist2['siglo'] = np.select(siglos, valores)

#ahora, confirmamos que las columnas en attacks_historicos y attacks_1_3 tienen las mismas columnas
atthist2.columns == attacks2.columns

### Ejercicio 4.
Es el momento de unir la información de todos los ataques de tiburón que acabamos de crear con el DataFrame de información adicional. Debatid que método es el más correcto para unir toda esta información y ejecutadlo.
📌 NOTA Cuando hagáis la unión considerad que la información que más nos interesa es la del DataFrame que creamos en el apartado anterior, lo cual determinará el tipo de unión.

        
        Fusionamos ambas bases (atthist2 y attacks2). Optamos por un merge() ya que es el método diseñado para combinar datos en columnas comunes:

nuevo_attack = attacks2.merge(atthist2, how='outer')
#sería lo mismo que hacer un .concat() pero especificando que se aplique en el axis[0], es decir, las filas.

nuevo_attack.info()

        Comprobamos las tres últimas columnas de nuestro nuevo DF (nuevo_attack), para confirmar que los tres últimos casos coinciden con las del df - atthist2. 
        
        En los scripts siguientes se comprueba que los ultimos case_number coinciden con los IDs de atthist2:

nuevo_attack.tail(3)

atthist2.tail(3)

    Fusión de nuevo_attacks con attacks_adicional. Optamos por concat() ya que nos interesa unir las columnas.

mega_nuevo_attack = pd.concat([nuevo_attack, attadic2], axis=1)

mega_nuevo_attack.columns

### Ejercicio 5.
Guardad el dataframe nuevo en un csv que usaremos mañana en la clase de pair programming.


mega_nuevo_attack.to_csv('../Archivos/attacks_1_4.csv', index=False)
