==> Pandas/6-pd-groupby.ipynb <==
import pandas as pd
import numpy as np

### Ejercicio 1.
¿Cuántos ataques hubo por país? Devuelve un DataFrame.
📌 NOTA Ordena los resultados de mayor a menor. Esto todavía no lo hemos visto, pero un DataFrame lo podemos ordenar con .sort_values(). Aquí tenéis una explicación de cómo hacerlo.

attacks = pd.read_csv('../Archivos/attacks_1_4.csv')
attacks.head(3)

attacks.columns

paises = attacks['country'].value_counts()
paises = paises.reset_index()
paises.sort_values(by='country', ascending=False)

### Ejercicio 2.
¿Cuántos ataques ha habido por año? Devuelve un DataFrame. De nuevo ordena los resultados de mayor a menor.

#añadimos el head() al final ya que de lo contrario nos enseña todas las filas independientemente de lo que le pidamos X desde el método .head()
#https://stackoverflow.com/a/36074520
attack_per_year = attacks.sort_values(['year'], ascending=False).groupby('year').head() 
attack_per_year.head(3)

### Ejercicio 3.
¿Cuántos ataques hubo por año y sexo? Devolved un DataFrame solo del año 2000 en adelante.

year2000 = attacks[attacks['year'] >= 2000]
ejercicio3 = year2000.groupby('year')['sex']

df_ejercicio3 = ejercicio3.count().reset_index()
df_ejercicio3

### Ejercicio 4.
Primero extraed solo las filas que correspondan a USA, AUSTRALIA y SOUTH AFRICA. ¿Cuántos de los ataques fueron letales por sexo? Devuelve un DataFrame.

lista_paises = ['USA', 'AUSTRALIA', 'SOUTH AFRICA']
attacks_por_pais = attacks[attacks['country'].isin(lista_paises)]

df_letal = attacks_por_pais[(attacks_por_pais['fatal_(y/n)'] == 'Y')]
df_letal.head(3)

df_letal_sexo = (df_letal.groupby('sex')['case_number'].count()).reset_index()
df_letal_sexo.head(3)

### Ejercicio 5.
Usando el DataFrame creado en el ejercicio anterior. ¿En qué país hubo más ataques?

ej5 = (attacks_por_pais.groupby('country')['case_number'].count()).reset_index()
ej5

print(f'La mayoría de ataques tuvo lugar en {ej5.index[2]}, con un total de {ej5.values[2]} ataques.')

### Ejercicio 6.
¿Qué y cuántos tipos de ataques hubo por país?

ej6 = pd.DataFrame(df_letal.groupby(['type', 'country'])['case_number'].count())
ej6
