==> Pandas/7-pd-apply.ipynb <==
import pandas as pd
import numpy as np
import regex as re

### Ejercicio 1
Cread una columna nueva y una función que nos de el mes en el que ocurrió el ataque. Tened en cuenta que no todas las filas tienen la misma estructura y que puede que no haya la información de mes. En ese caso devolved un nulo (NaN).

pd7 = pd.read_csv("../Archivos/attacks_1_4.csv")
pd7.head(3)

pd7.columns

pd7_copia = pd7.copy()

def sacar_mes(x):
    try:
        return re.findall("([a-zA-Z]{1}[a-z]{2})\-", x)[0]
    except:
        return np.nan #necesitamos añadir los valores nulos a la función, ya que de lo contrario, nos da error

pd7_copia["mes"] = pd7_copia['date'].apply(sacar_mes)
pd7_copia['mes']

### Ejercicio 2
Vamos con la columna fatal_(y/n). Lo primero que tenéis que hacer es evaluar los valores únicos que hay. Esperaríamos tener solo dos valores: y y n, pero tristemente no será así 😭. Cread una columna nueva y una función que devuelva únicamente y o n y que devuelva un nulo (NaN) si no se encuentra el valor.

#valoramos los casos unicos.
pd7_copia["fatal_(y/n)"].unique()

def limpiar (col):
    try:
        if col == ' N' or col == 'N ' or col == 'N':
            return 'N'
        elif col == 'y' or col == 'Y':
            return 'Y'
        elif col == 'Y':
            return 'Y'
    except:
        return np.nan

pd7_copia['new_fatal_(y/n)'] = pd7_copia['fatal_(y/n)'].apply(limpiar)

pd7_copia['new_fatal_(y/n)'].unique()

### Ejercicio 3.
Seguid la misma lógica que en el ejercicio anterior pero con la columna sex. Pero en este caso vamos a incluir un retito más: En este caso querremos tener dos valores únicos F y M. De nuevo en la columna hay otros carácteres raros como 'M ', 'lli', 'N', '.'. En caso de que sea nulo, que la función devuelva nulo (np.nan).

#extraemos los valores únicos de la columna 'sex':
pd7_copia['sex'].unique()

#creamos la función:
def norm_sex (col):
    try:
        if col == 'F':
            return 'F'
        elif col == 'M':
            return 'M'
    except:
        return np.nan

pd7_copia['sex_norm'] = pd7['sex'].apply(norm_sex)
#comprobamos los valores únicos de la nueva columna:
pd7_copia['sex_norm'].unique()

### Ejercicio 4.
Una vez que hayáis terminado eliminad las columnas originales conservando las nuevas.

pd7_copia = pd7_copia.drop(['sex', 'fatal_(y/n)'], axis=1)

pd7_copia.columns

### Ejercicio 5.
Guardad el dataframe en un csv que usaremos más adelante

pd7_copia.to_csv('../Archivos/attacks_def.csv')
